REFACTOR: split elements.py out into separate files.
