+++ src/md_html/handler.py
@@ -52,6 +52,7 @@ parser.add_argument(
     help="Print the Version number and exit.",
 )
 
+
 def usage() -> None:
     """Print usage documentation and exit."""
     parser.print_usage()
