+++ tests/elements/test_base_element.py
@@ -2,11 +2,9 @@
 
 # ruff: noqa: S101 S301
 
-import json
-
 import pytest
 
-from md_html.elements import HTML, HTMLBaseClass, P
+from md_html.elements.base_element import HTMLBaseClass
 from md_html.exceptions import HTMLParseError
 from md_html.utils import Elements
 
@@ -120,72 +118,3 @@ class TestHTMLBaseClass:
         test_instance.kind = kind
         test_instance.attr = attr
         assert test_instance.html == expected
-
-
-class TestParagraphElements:
-    """Test the methods of the Paragraph class."""
-
-    @pytest.mark.parametrize(
-        "content,expected",
-        [
-            ("Hello World!", "<p>Hello World!</p>"),
-            ("Hello <div>", "<p>Hello &lt;div&gt;</p>"),
-            ("<p>", "<p>&lt;p&gt;</p>"),
-            ("**strong text**", "<p>**strong text**</p>"),
-            ("*emphasis text*", "<p>*emphasis text*</p>"),
-            (
-                "**strong text with *emphasis***",
-                "<p>**strong text with *emphasis***</p>",
-            ),
-            (
-                "***strong text with emphasis***",
-                "<p>***strong text with emphasis***</p>",
-            ),
-        ],
-    )
-    def test_html_render(self, content, expected) -> None:
-        """R-BICEP: Right."""
-        test = P(content)
-        assert test.html == expected
-
-
-class TestHTMLElements:
-    """Test the methods of the HTML class."""
-
-    @pytest.mark.parametrize(
-        "title,file,attr",
-        [
-            ("Markdown Document", "table", []),
-            ("Markdown Document", "inline-img", []),
-            ("Markdown Document", "quantum-slopes", []),
-            ("Markdown Document", "blockquote", []),
-            ("Markdown Document", "perfectly-ripe-tomatoes", []),
-            ("Markdown Document", "a-grand-day-out", []),
-            ("Markdown Document", "a-tower-of-blocks", []),
-        ],
-    )
-    def test_html_render_of_html_elements(
-        self, title, file, attr, snapshot
-    ) -> None:
-        """R-BICEP: Right."""
-        snapshot.snapshot_dir = "tests/snapshots/html"
-        with open(f"./tests/data/sections/{file}.json", "rb") as _file:
-            sections = json.load(_file)
-        test = HTML(title=title, sections=sections, atr=attr)
-        snapshot.assert_match(test.html, f"{file}.html")
-
-    @pytest.mark.parametrize(
-        "title,sections,attr,expected",
-        [
-            (
-                "Markdown Document",
-                [],
-                ["class='light'"],
-                "<html lang='en_gb' dir='ltr' class='light'>",
-            ),
-        ],
-    )
-    def test_html_can_parse_custom_attr(self, title, sections, attr, expected) -> None:
-        """R-BICEP: Right."""
-        test = HTML(title=title, sections=sections, attr=attr)
-        assert test.html.split("\n")[0] == expected
